Gen2 firmware: Checking in to eliminate fear of losing a large bit of work.  Gen2 firmware has been updated to add a new command to set all incandescent bulbs as either on or off in a single command.  (Note:  This overwrites all other setting such as blink and fade).  It makes reduces amount of MPF commands to update incandescent bulbs in half.  Update interface document to add new command.
MPF:
Initial MPF implementation for OPP hardware.  All I can say is that it doesn't crash when the test machine boots.  (If you start it, it will crash)  More testing will happen in the next few days.  Added MPFTest machine to match board sent to Brian so he can play with it.  The MPFUpdates directory contains the files that need to be added or merged with the MPF python code.

MPF/SS3:
Added an attract light show and some sounds.