Firmware version 0.2.0.1: Added RS232I_UPGRADE_OTHER_BRD (0x18) command.  Make RS232 port transparent so other board can be upgraded.  65 EOM characters will end and reset the card.
Gen2Test.py:
Add support for -upgrade flag.  When set, ../../Creator/Gen2Images directory is examined looking for newest version of the firmware.  If there is a new version of firmware, it will upgrade each card one at a time, and rewrite the serial number into the flash (preserving it).